Revert "Seggrigate API fixtures in their own component fixture file, categorized into core and component packages (#9067)"
This reverts commit ecbf6f7fe7f348597f8f619ee404106cd8cf0788.
