+++ conftest.py
@@ -14,37 +14,20 @@ pytest_plugins = [
     'pytest_plugins.settings_skip',
     'pytest_plugins.rerun_rp.rerun_rp',
     # Fixtures
-    'pytest_fixtures.core.broker',
-    'pytest_fixtures.core.contenthosts',
-    'pytest_fixtures.core.reporting',
-    'pytest_fixtures.core.sys',
-    'pytest_fixtures.core.upgrade',
-    'pytest_fixtures.core.xdist',
+    'pytest_fixtures.api_fixtures',
+    'pytest_fixtures.broker',
+    'pytest_fixtures.content_hosts',
+    'pytest_fixtures.reporting_fixtures',
+    'pytest_fixtures.xdist',
     # Component Fixtures
-    'pytest_fixtures.component.activationkey',
-    'pytest_fixtures.component.ansible',
-    'pytest_fixtures.component.architecture',
-    'pytest_fixtures.component.computeprofile',
-    'pytest_fixtures.component.contentview',
-    'pytest_fixtures.component.domain',
-    'pytest_fixtures.component.host',
-    'pytest_fixtures.component.hostgroup',
-    'pytest_fixtures.component.lce',
-    'pytest_fixtures.component.os',
-    'pytest_fixtures.component.oscap',
-    'pytest_fixtures.component.provision_azure',
-    'pytest_fixtures.component.provision_gce',
-    'pytest_fixtures.component.provision_libvirt',
-    'pytest_fixtures.component.provisioning_template',
-    'pytest_fixtures.component.puppet',
-    'pytest_fixtures.component.repository',
-    'pytest_fixtures.component.rh_cloud',
-    'pytest_fixtures.component.satellite_auth',
-    'pytest_fixtures.component.settings',
-    'pytest_fixtures.component.smartproxy',
-    'pytest_fixtures.component.subnet',
-    'pytest_fixtures.component.subscription',
-    'pytest_fixtures.component.taxonomy',
-    'pytest_fixtures.component.templatesync',
-    'pytest_fixtures.component.user_role',
+    'pytest_fixtures.ansible_fixtures',
+    'pytest_fixtures.oscap_fixtures',
+    'pytest_fixtures.rh_cloud',
+    'pytest_fixtures.satellite_auth',
+    'pytest_fixtures.sys_fixtures',
+    'pytest_fixtures.subscription_fixtures',
+    'pytest_fixtures.smartproxy_fixtures',
+    'pytest_fixtures.templatesync_fixtures',
+    'pytest_fixtures.user_fixtures',
+    'pytest_fixtures.upgrade_fixtures',
 ]
