--- pytest_fixtures/component/computeprofile.py
@@ -1,8 +1,0 @@
-# Compute Profile Fixtures
-import pytest
-from nailgun import entities
-
-
-@pytest.fixture(scope='module')
-def module_compute_profile():
-    return entities.ComputeProfile().create()
