+++ pytest_fixtures/satellite_auth.py
