+++ pytest_plugins/fixture_markers.py
@@ -25,9 +25,9 @@ def pytest_configure(config):
 
 
 def pytest_collection_modifyitems(session, items, config):
-    from pytest_fixtures.core import contenthosts
+    from pytest_fixtures import content_hosts
 
-    content_host_fixture_names = [m[0] for m in getmembers(contenthosts, isfunction)]
+    content_host_fixture_names = [m[0] for m in getmembers(content_hosts, isfunction)]
     for item in items:
         if set(item.fixturenames).intersection(set(content_host_fixture_names)):
             # TODO check param for indirect version parametrization
